requirements.txt
django-bower==5.2.0
django-compressor==2.1.1
django-sass-processor==0.5.4
libsass==0.13.2
